added midi control of the center frequency

+++ Source/PluginProcessor.h
@@ -63,7 +63,7 @@ public:
 	const String PID_Q = "q";
 	const String PID_GAIN = "gain";
 
-	int midiDebugNumber_;
+	float midiDebugNumber_;
 private:
 	MyBandPass** wahFilters_;
 	int numWahFilters_;
